+++ com.unity.formats.alembic/Editor/UnityRecorder/Input/AlembicInput.cs
@@ -1,0 +1,10 @@
+#if RECORDER_AVAILABLE
+using UnityEditor.Recorder;
+
+namespace UnityEditor.Formats.Alembic.Recorder
+{
+    class AlembicInput : RecorderInput
+    {
+    }
+}
+#endif