+++ com.unity.formats.alembic/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Changes
+- Added Unity recorder integration (compatible with Unity Recorder >= 2.2.0).
 
 ## [2.0.1-preview.1] - 2019-08-07
 ### Changes